Update URL request formation to reflect changes in API v3.
API v3 expects all POST and PATCH calls to send data as JSON rather than attached to the query string. First we serialise the incoming params dictionary into JSON with NSJSONSerialization, then set the Content-Type of our request to application/json. Then we set the body of the request to the serialised JSON before sending it. Note that this requires a bit of a workaround at the moment due to half the engine using v2 and half using v3. This will be simplified once everything's on v3.