Make -sendRequest's return type variable.
Still returns an id, but this can be either an NSArray, an NSNumber with YES or NO for boolean handling, or an NSError.
- If it's an array, json was received and parsed as expected.
- If it's an NSNumber YES, that's a successful call to a method that expects a No Content response.
- If it's an NSNumber NO, that's a successful call to a method that returns an expected 404.
- If it's an NSError, it's either a connection error, an HTTP error other than the above, or a parser error. NSError instance can be inspected to determine which.

This all enables us to return e.g. a BOOL from a call to UAGE. This is useful in situations where Github will return no data, when calling e.g. -follows:completion:. This can now return a simple YES or NO. Error handling is also significantly improved, in that it now exists.