Trying out new approach on -respositoriesWithCompletion:failure:.
Now we tell UAGE to invoke the necessary API call and return the result.
In -invoke, we create the NSError and id that get returned when the connection either succeeds or fails. This avoids replication of code all over the blinking place, and keeps each individual method down to a single line.
(Also gives us the option to completely revise the way this works without having to change every single method again. Can just change the NSInvocation.)