Error passing now *finally* works properly.
When creating an invocation, we have to create and assign both an NSError and a pointer to that NSError. We then send the address of the pointer to the error to the invocation by reference. Phew.
(Also we now return after we call our failureBlock, otherwise we'll call the successBlock afterwards. And that ain't right.)