Fixed crash caused by error handling in -invoke:… Yes, again.
Method signatures differ between invocations, but the last argument is always where the NSError lives. Previous implementation used '5' hard-coded as the argument for the error, but this (obviously) caused a crash when using a method signature that had the error at a different position.
Now, we examine the method signature, find out the number of arguments, subtract 1 as it's zero-based, and put the error there.